src/requirements.txt
fastapi==0.75.0
uvicorn==0.15.0

#for database
sqlalchemy==1.4.32

#hashing
passlib==1.7.4
bcrypt==3.2.0

#for creating jwt tokens
python-jose==3.3.0

pydantic==1.9.0

